requirements.txt
pyrogram
tgCrypto
python-dotenv
requests
